pyproject.toml
[tool.poetry]
name = "nendo-li"
version = "0.1.0"
description = "Nendoroid web DB project"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1"
Pillow = "^9.2.0"
djangorestframework = "^3.13.1"
django-cors-headers = "^3.13.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
